change phone nr data type -- new changeset

+++ src/main/java/Wr40/cardiary/model/entity/User.java
@@ -22,7 +22,7 @@ public class User {
     private String password;
     @Column(length = 45)
     private String email;
-    private short phoneNumber;
+    private Long phoneNumber;
     @CreationTimestamp
     private LocalDateTime creationDate;
     @OneToMany
